clean up folder structure
