+++ assembly/lmdGhost/index.ts
@@ -1,0 +1,2 @@
+export * from "./interface";
+export * from "./statefulDag";
