+++ assembly/lmdGhost/statefulDag/index.ts
@@ -1,0 +1,2 @@
+export * from "./lmdGhost";
+export * from "./attestationAggregator";
