+++ assembly/types/index.ts
@@ -1,0 +1,5 @@
+export type Root = string;
+export type bytes32 = Buffer;
+export type Gwei = u64;
+export type Slot = u32;
+export type ValidatorIndex = u32;
