+++ package.json
@@ -96,7 +96,7 @@
     "@types/ws": "^6.0.1",
     "@typescript-eslint/eslint-plugin": "^1.3.0",
     "@typescript-eslint/parser": "^1.3.0",
-    "assemblyscript": "AssemblyScript/assemblyscript",
+    "assemblyscript": "git://github.com/AssemblyScript/assemblyscript.git#dev",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "codecov": "^3.1.0",
